+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Fragments can now be used to fill repair kits (1/8 of a unit per) [#311]
 - Config to disable material mixing in the crafting grid when crafting parts. This is left enabled (old behavior by default).
 ### Fixed
 - Elytra not working on back slot with Caelus API [#302]